+++ apps/dashboard/app/entry.server.tsx
@@ -1,6 +1,6 @@
 import type { EntryContext } from "@remix-run/node";
 import * as Sentry from "@sentry/remix";
-import { initSentry } from "./shared/sentry";
+import { initSentry } from "@webstudio-is/remix";
 import { prisma } from "@webstudio-is/prisma-client";
 import { handleRequest } from "./shared/remix";
