+++ apps/designer/app/canvas/shared/sync/use-sync.ts
@@ -3,7 +3,7 @@ import { sync } from "immerhin";
 import { enqueue } from "./queue";
 import type { Build } from "@webstudio-is/project";
 import { Tree } from "@webstudio-is/react-sdk";
-import { restPatchPath } from "~/shared/router-utils";
+import { restPatchPath } from "~/shared/build-utils";
 
 export const useSync = ({
   treeId,
