+++ apps/designer/app/routes/auth/github/callback.tsx
@@ -1,6 +1,6 @@
 import { type LoaderArgs, redirect } from "@remix-run/node";
 import { authenticator } from "~/services/auth.server";
-import { loginPath } from "~/shared/router-utils";
+import { loginPath } from "~/shared/build-utils";
 import { AUTH_PROVIDERS } from "~/shared/session";
 import { returnToPath } from "~/services/cookie.server";
