+++ apps/designer/app/routes/login/index.tsx
@@ -5,7 +5,7 @@ import env from "~/env.server";
 
 import { Login, links } from "~/auth";
 import { useLoginErrorMessage } from "~/shared/session";
-import { dashboardPath } from "~/shared/router-utils";
+import { dashboardPath } from "~/shared/build-utils";
 import { returnToCookie } from "~/services/cookie.server";
 
 export { links };
