+++ packages/remix/src/index.ts
@@ -1,1 +1,5 @@
 export { default as configVercel } from "./config-vercel";
+export * from "./sentry";
+export * from "./handle-request.server";
+export * from "./root";
+export * as env from "./env";
