+++ packages/remix/src/path-utils.ts
@@ -1,8 +1,5 @@
 import { AUTH_PROVIDERS } from "~/shared/session";
-import type { Page, Project } from "@webstudio-is/project";
-import type { BuildMode } from "./build-params";
-import type { ThemeSetting } from "~/shared/theme";
-import env from "~/shared/env";
+import type { ThemeSetting } from "./theme";
 
 const searchParams = (params: Record<string, string | undefined | null>) => {
   const searchParams = new URLSearchParams();
@@ -57,30 +54,3 @@ export const restThemePath = ({ setting }: { setting: ThemeSetting }) =>
 export const restPatchPath = () => "/rest/patch";
 
 export const restPublishPath = () => "/rest/publish";
-
-export const getBuildUrl = ({
-  buildOrigin,
-  project,
-  page,
-  mode,
-}: {
-  buildOrigin: string;
-  project: Project;
-  page: Page;
-  mode?: BuildMode;
-}) => {
-  const url = new URL(buildOrigin);
-  url.pathname = page.path;
-
-  if (env.BUILD_REQUIRE_SUBDOMAIN) {
-    url.host = `${project.domain}.${url.host}`;
-  } else {
-    url.searchParams.set("projectId", project.id);
-  }
-
-  if (mode) {
-    url.searchParams.set("mode", mode);
-  }
-
-  return url.toString();
-};
